requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-experimental>=0.0.60
langchain-groq>=0.1.5

python-dotenv>=1.0.1
streamlit>=1.32.0

tiktoken>=0.7.0
faiss-cpu>=1.7.4

protobuf>=4.25.3

pymysql>=1.1.0
mysql-connector-python>=8.3.0

sentence-transformers>=2.6.1
chromadb>=0.4.24